+++ packages/cf-deploy-config-sub-generator/src/app/index.ts
@@ -41,10 +41,10 @@ import { CfDeployConfigOptions } from './types';
 import {
     type CfAppRouterDeployConfigAnswers,
     type CfDeployConfigQuestions,
+    type CfDeployConfigPromptOptions,
     CfDeployConfigAnswers,
     RouterModuleType
 } from '@sap-ux/cf-deploy-config-inquirer';
-import type { YeomanEnvironment } from '@sap-ux/fiori-generator-shared';
 import type { Answers } from 'inquirer';
 
 /**
@@ -102,10 +102,6 @@ export default class extends DeploymentGenerator {
     public async initializing(): Promise<void> {
         await super.initializing();
         await initI18n();
-
-        if ((this.env as unknown as YeomanEnvironment).conflicter) {
-            (this.env as unknown as YeomanEnvironment).conflicter.force = this.options.force ?? true;
-        }
 
         DeploymentGenerator.logger?.debug(t('cfGen.debug.initTelemetry'));
 
@@ -403,4 +399,11 @@ export default class extends DeploymentGenerator {
 
 export { getCFQuestions, loadManifest };
 export { API_BUSINESS_HUB_ENTERPRISE_PREFIX, DESTINATION_AUTHTYPE_NOTFOUND };
-export { CfDeployConfigOptions, CfDeployConfigAnswers, CfDeployConfigQuestions, ApiHubConfig, ApiHubType };
+export {
+    CfDeployConfigOptions,
+    CfDeployConfigAnswers,
+    CfDeployConfigQuestions,
+    CfDeployConfigPromptOptions,
+    ApiHubConfig,
+    ApiHubType
+};