+++ packages/fe-fpm-writer/test/unit/page/custom.test.ts
@@ -369,4 +369,19 @@ describe('CustomPage', () => {
         //check
         expect((fs.readJSON(join(target, 'webapp/manifest.json')) as any)?.['sap.ui5'].dependencies).toMatchSnapshot();
     });
+
+    test('Add when "sap.fe.ariba" dependency is listed', async () => {
+        const testManifest = JSON.parse(testAppManifest);
+        testManifest['sap.ui5'].dependencies.libs['sap.fe.ariba'] = {};
+        const minimalInput: CustomPage = {
+            name: 'CustomPage',
+            entity: 'RootEntity'
+        };
+        const target = join(testDir, 'ariba');
+        fs.write(join(target, 'webapp/manifest.json'), JSON.stringify(testManifest));
+        //act
+        await generateCustomPage(target, minimalInput, fs);
+        //check
+        expect(fs.readJSON(join(target, 'webapp/manifest.json'))).toMatchSnapshot();
+    });
 });