+++ packages/fiori-app-sub-generator/test/int/fiori-elements/writing-v2.test.ts
@@ -89,7 +89,21 @@ describe('Generate v2 apps', () => {
             }) as Project,
             floorplan: FloorplanFE.FE_LROP,
             service: destService,
-            entityRelatedConfig: v2EntityConfig
+            entityRelatedConfig: v2EntityConfig,
+            appGenInfo: {
+                externalParameters: {
+                    'testAdditionalEntry': [
+                        {
+                            key: 'testKey1',
+                            value: 'testValue1'
+                        },
+                        {
+                            key: 'testKey2',
+                            value: 'testValue2'
+                        }
+                    ]
+                }
+            }
         });
         await runWritingPhaseGen(testState);
         expect(join(testDir, testProjectName)).toMatchFolder(getExpectedOutputPath(testProjectName), ignoreMatcherOpts);