+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,116 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 0.1.14
+
+### Patch Changes
+
+-   Updated dependencies [95a816d]
+    -   @sap-ux/fiori-generator-shared@0.12.6
+    -   @sap-ux/flp-config-inquirer@0.3.14
+    -   @sap-ux/inquirer-common@0.7.7
+    -   @sap-ux/adp-tooling@0.14.14
+
+## 0.1.13
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+    -   @sap-ux/flp-config-inquirer@0.3.13
+    -   @sap-ux/inquirer-common@0.7.6
+    -   @sap-ux/adp-tooling@0.14.13
+
+## 0.1.12
+
+### Patch Changes
+
+-   Updated dependencies [61d4060]
+    -   @sap-ux/axios-extension@1.21.2
+    -   @sap-ux/adp-tooling@0.14.12
+    -   @sap-ux/system-access@0.6.2
+    -   @sap-ux/flp-config-inquirer@0.3.12
+
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies [91726b0]
+    -   @sap-ux/fiori-generator-shared@0.12.4
+    -   @sap-ux/flp-config-inquirer@0.3.11
+    -   @sap-ux/inquirer-common@0.7.5
+    -   @sap-ux/adp-tooling@0.14.11
+
+## 0.1.10
+
+### Patch Changes
+
+-   Updated dependencies [61ea5c0]
+    -   @sap-ux/adp-tooling@0.14.10
+    -   @sap-ux/flp-config-inquirer@0.3.10
+    -   @sap-ux/project-access@1.30.2
+    -   @sap-ux/axios-extension@1.21.1
+    -   @sap-ux/fiori-generator-shared@0.12.3
+    -   @sap-ux/system-access@0.6.1
+    -   @sap-ux/inquirer-common@0.7.4
+
+## 0.1.9
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.9
+-   @sap-ux/flp-config-inquirer@0.3.9
+
+## 0.1.8
+
+### Patch Changes
+
+-   Updated dependencies [ac55cca]
+    -   @sap-ux/fiori-generator-shared@0.12.2
+    -   @sap-ux/flp-config-inquirer@0.3.8
+    -   @sap-ux/inquirer-common@0.7.3
+    -   @sap-ux/adp-tooling@0.14.8
+
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [b3fe5b8]
+    -   @sap-ux/inquirer-common@0.7.2
+    -   @sap-ux/adp-tooling@0.14.7
+    -   @sap-ux/flp-config-inquirer@0.3.7
+
+## 0.1.6
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.6
+-   @sap-ux/flp-config-inquirer@0.3.6
+
+## 0.1.5
+
+### Patch Changes
+
+-   Updated dependencies [2224d63]
+    -   @sap-ux/axios-extension@1.21.1
+    -   @sap-ux/adp-tooling@0.14.5
+    -   @sap-ux/system-access@0.6.1
+    -   @sap-ux/flp-config-inquirer@0.3.5
+
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [8e44d02]
+    -   @sap-ux/adp-tooling@0.14.4
+    -   @sap-ux/flp-config-inquirer@0.3.4
+
+## 0.1.3
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.3
+-   @sap-ux/flp-config-inquirer@0.3.3
+
 ## 0.1.2
 
 ### Patch Changes