+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,71 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.8
+
+### Patch Changes
+
+-   3630ec9: fixs for abap headless
+
+## 0.2.7
+
+### Patch Changes
+
+-   Updated dependencies [95a816d]
+    -   @sap-ux/fiori-generator-shared@0.12.6
+    -   @sap-ux/deploy-config-generator-shared@0.1.6
+    -   @sap-ux/inquirer-common@0.7.7
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.7
+
+## 0.2.6
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+    -   @sap-ux/deploy-config-generator-shared@0.1.5
+    -   @sap-ux/inquirer-common@0.7.6
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.6
+
+## 0.2.5
+
+### Patch Changes
+
+-   Updated dependencies [91726b0]
+    -   @sap-ux/fiori-generator-shared@0.12.4
+    -   @sap-ux/deploy-config-generator-shared@0.1.4
+    -   @sap-ux/inquirer-common@0.7.5
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.5
+
+## 0.2.4
+
+### Patch Changes
+
+-   @sap-ux/cf-deploy-config-writer@0.3.2
+-   @sap-ux/project-access@1.30.2
+-   @sap-ux/fiori-generator-shared@0.12.3
+-   @sap-ux/inquirer-common@0.7.4
+-   @sap-ux/deploy-config-generator-shared@0.1.3
+-   @sap-ux/cf-deploy-config-inquirer@0.3.4
+
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [ac55cca]
+    -   @sap-ux/fiori-generator-shared@0.12.2
+    -   @sap-ux/deploy-config-generator-shared@0.1.2
+    -   @sap-ux/inquirer-common@0.7.3
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.3
+
+## 0.2.2
+
+### Patch Changes
+
+-   Updated dependencies [b3fe5b8]
+    -   @sap-ux/inquirer-common@0.7.2
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.2
+    -   @sap-ux/deploy-config-generator-shared@0.1.1
+
 ## 0.2.1
 
 ### Patch Changes