+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.35.0
+
+### Minor Changes
+
+-   ccd85c8: Support 'sap.fe.ariba' dependency for new pages'. If the dependency is listed in the manifest, it will be used as the template(property "name") for new LR and OP pages.
+
+## 0.34.3
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.2
+-   @sap-ux/fiori-annotation-api@0.5.7
+
+## 0.34.2
+
+### Patch Changes
+
+-   Updated dependencies [e605d30]
+    -   @sap-ux/fiori-annotation-api@0.5.6
+
 ## 0.34.1
 
 ### Patch Changes