+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.14.5
+
+### Patch Changes
+
+-   920c23d: Fix custom connector for flex changes in UI5 < 1.78
+
+## 0.14.4
+
+### Patch Changes
+
+-   225e7d7: fix: Add message when controller extension pending change is created
+
 ## 0.14.3
 
 ### Patch Changes