+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/system-access
 
+## 0.6.2
+
+### Patch Changes
+
+-   Updated dependencies [61d4060]
+    -   @sap-ux/axios-extension@1.21.2
+
+## 0.6.1
+
+### Patch Changes
+
+-   Updated dependencies [2224d63]
+    -   @sap-ux/axios-extension@1.21.1
+
 ## 0.6.0
 
 ### Minor Changes